+++ package.json
@@ -2,9 +2,10 @@
 	"name": "tlk",
 	"version": "3.1.0",
 	"description": "Free video call app for the web.",
-	"main": "server.js",
+	"main": "index.js",
 	"scripts": {
 		"test": "echo \"Error: no test specified\" && exit 1",
+		"start": "node index.js",
 		"postinstall": "node postinstall.js"
 	},
 	"repository": {
